+++ src/main/java/terramine/common/item/curio/hands/BandOfRegenerationItem.java
@@ -2,7 +2,9 @@ package terramine.common.item.curio.hands;
 
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
+import terramine.common.init.ModItems;
 import terramine.common.item.curio.TrinketTerrariaItem;
+import terramine.common.trinkets.TrinketsHelper;
 
 public class BandOfRegenerationItem extends TrinketTerrariaItem {
 
@@ -17,7 +19,7 @@ public class BandOfRegenerationItem extends TrinketTerrariaItem {
 
 	@Override
 	public void curioTick(LivingEntity player, ItemStack stack) {
-		if (player != null) {
+		if (player != null && !TrinketsHelper.isEquipped(ModItems.CHARM_OF_MYTHS, player)) {
 			timer += 1;
 			if (timer >= 50) {
 				player.heal(0.5f);